Add files via upload
Got done making the code, I decided to make it more terminal based as supposed to adding collision boxes in window(I was getting bored), If i had to revisit this i would definately want to add an idle animation, sorting, animations for searching & deletion, but for now, it works, and the update and draw functions are part of every node so some functionality of linked list is being used. I also created the dragon head and dragon body sprites, more like i just took a ss of the one i liked online and traced it and converted it to points. Couldve just used sprites but i wanted to use trignometry to rotate the points myself :p well Will upload the fully finished version in a bit after im done making its documentation and explanation.